on every result item, temp file is set to undefined at the start

+++ tsnuxt.sh
@@ -1,11 +1,12 @@
 #!/bin/bash
 
 tsdir="/home/uros/Documents/programiranje/vue/ui-lishuuro/node_modules/typescript/lib"
+
 oldstart="var ts ="
 newstart='const tempp = require("../lib/temp");'
 
 oldexpands="withoutExpandos : filteredDeclarations;"
-newexpands="tempp.nuxtComp(results, node); tempp.autoImport(results, node); "
+newexpands="tempp.nuxtComp(results, node);"
 
 oldcreatedef="(function createDefinitionInfoFromName\(.*\).*\n.*getSourceFile.*;)"
 
@@ -15,15 +16,18 @@ oldTextSpan="(textSpan = createTextSpanFromNode\(name, sourceFile\);)"
 
 tsfile="$tsdir/typescript.js"
 
-echo "$tsfile"
+echo $1
 
-cp -r "$tsfile" "$tsdir/original-typescript.js"
+[ "1" -eq "$1" ] && cp -r "$tsdir/original-typescript.js" "$tsdir/typescript.js" && exit 1
+
+test ! -f "$tsdir/original-typescript.js" && cp -r "$tsfile" "$tsdir/original-typescript.js"
 cp -r temp.js "$tsdir"
 sd "$oldstart" "$newstart\n$oldstart" "$tsfile"
 sd "$oldexpands" "$oldexpands\n  $newexpands\n" "$tsfile"
 sd "$oldcreatedef" '$1
     let tfn = sourceFile.tempFileName;
-    let fileName = tfn ? tfn : sourceFile.fileName;' "$tsfile"
+    let fileName = tfn ? tfn : sourceFile.fileName;
+    tempp.log2(`tfn: ${fileName}`);' "$tsfile"
 
 sd -f ms "(function createDefinitionInfoFromName\(.*?fileName:).*?,(.*?;.*?\})" '$1 fileName, $2' "$tsfile"
 sd -f ms "$oldTextSpan" '
@@ -35,3 +39,5 @@ sd -f ms "$oldTextSpan" '
 
 sd "$olddefssymbol" '$1
     if(defs.length > 0) { return defs }' "$tsfile"
+
+echo "done"
